--- a/logger.js
+++ b/logger.js
@@ -10,10 +10,13 @@ function logger (opts, stream) {
     wrap(ctx.req, ctx.res)
     ctx.log = ctx.request.log = ctx.response.log = ctx.req.log
     return next().catch(function (err) {
-      ctx.log.error({ err })
+      if (!('err' in ctx.res)) {
+        ctx.res.err = err
+      }
       throw err
     })
   }
+
   pino.logger = wrap.logger
   return pino
 }