- remove explicit log statement in favor of providing res.err and
  err.status as recognized by pino-http
- added pino ^5.12.0 as a devDependency to support extreme mode
  benchmark

fixes #19