@jstewmon 2019-04-04
Leaving a note for myself that I didn’t confirm a test case exists for ensuring the status is correctly captured when using ctx.throw