@jstewmon 2019-04-04
As it turns out, we only needed to assign `ctx.res.err`.

The Error instances created by http-errors assigns both `status` and `statusCode` as enumerable props, so the value is logged under two keys, which I don't love. But, the only way I can think to address it here would be to either delete one of them or redefine one of them as non-enumerable, which of course could violate someone's expectations elsewhere.

Is it possible to customize the err serializer to pick `statusCode` and omit `status` from the keys that are added to the JSON? Seems like that should be possible, but it's been a while since I worked with a custom serializer.