@amckibben 2019-12-04
Yeah ... this being so outdated is (I think) causing the pino `redact` option not to work / be available which is a bummer.